fix(server): shoot with float damage

+++ shared/ecs/Component.hpp
@@ -42,6 +42,6 @@ namespace rtype {
 
     struct Enemy {
         int damage;
-        int life;
+        float life;
     };
 }
